+++ src/app/api/post/[id]/route.ts
@@ -1,0 +1,21 @@
+import { cookies } from 'next/headers';
+import { NextResponse } from 'next/server';
+
+export async function GET(
+  req: Request,
+  { params }: { params: Promise<{ id: string }> },
+) {
+  const { id } = await params;
+  const token = cookies().get('token')?.value;
+
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/post/${id}`, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+      ...(token ? { Authorization: `Bearer ${token}` } : {}),
+    },
+  });
+
+  const data = await res.json();
+  return NextResponse.json(data);
+}
